# src/main/resources/openapi/api.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: API Quick Approval System
  description: Api для взаимодействия с базой данных
  contact:
    name: Kirill & Igor
...
servers:
  - url: localhost:8080/api/v1
paths:
  /user:
    get:
      tags:
        - user
      description: данные по всем пользователям
      operationId: allUsers
      responses:
        '200':
          description: user without password response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserWithoutPassword'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - user
      description: создать нового пользователя
      operationId: addUser
      requestBody:
        description: пользователь, которого нужно добавить
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: user response
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/{id}:
    get:
      tags:
        - user
      description: данные по пользователю
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: id пользователя, которого нужно получить
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: user response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithoutPassword'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
        - user
      description: обновить данные пользователя
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: id пользователя, которого нужно обновить
          required: true
          schema:
            type: integer
      requestBody:
        description: новые данные для пользователя
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: user response
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/{id}/role:
    get:
      tags:
        - user
      description: список всех ролей пользователя
      operationId: getRoleListOfUserById
      parameters:
        - name: id
          in: path
          description: id пользователя, список ролей которого нужно получить
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: role response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/{id}/task:
    get:
      tags:
        - user
      description: список всех задач пользователя
      operationId: getTaskListOfUserById
      parameters:
        - name: id
          in: path
          description: id пользователя, список задач которого нужно получить
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: task response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/{id}/task/status/active:
    get:
      tags:
        - user
      description: список задач пользователя со статусами {active; sended}
      operationId: getActiveTaskListOfUserById
      parameters:
        - name: id
          in: path
          description: id пользователя, список задач которого нужно получить
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: task response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/{id}/task/status/wait:
    get:
      tags:
        - user
      description: список задач пользователя со статусом {wait}
      operationId: getWaitTaskListOfUserById
      parameters:
        - name: id
          in: path
          description: id пользователя, список задач которого нужно получить
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: task response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /user/{id}/task/status/complete:
    get:
      tags:
        - user
      description: список задач пользователя со статусами {agreed, denied}
      operationId: getCompleteTaskListOfUserById
      parameters:
        - name: id
          in: path
          description: id пользователя, список задач которого нужно получить
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: task response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /user/role/{id}:
    post:
      tags:
        - user
      description: добавить пользователю роль
      operationId: addRoleToUserById
      parameters:
        - name: id
          in: path
          description: id пользователя, которому нужно добавить роль
          required: true
          schema:
            type: integer
      requestBody:
        description: новая роль для пользователя
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: role response
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      type: object
      required:
        - id_user
        - fio
        - login
        - password
        - email
        - telegram_chat_id
      properties:
        id_user:
          type: integer
        fio:
          type: string
        login:
          type: string
        password:
          type: string
        email:
          type: string
        telegram_chat_id:
          type: integer
    UserWithoutPassword:
      type: object
      required:
        - id_user
        - fio
        - login
        - email
        - telegram_chat_id
      properties:
        id_user:
          type: integer
        fio:
          type: string
        login:
          type: string
        email:
          type: string
        telegram_chat_id:
          type: integer
    Role:
      type: object
      required:
        - id_role
        - name
      properties:
        id_role:
          type: integer
        name:
          type: string
    Task:
      type: object
      required:
        - id_task
        - process_id
        - user_performer_id
        - role_performer_id
        - date_start
        - date_end_planning
        - date_end_fact
        - status_id
      properties:
        id_task:
          type: integer
        process_id:
          type: integer
        user_performer_id:
          type: integer
        role_performer_id:
          type: integer
        date_start:
          $ref: '#/components/schemas/Timestamp'
        date_end_planning:
          $ref: '#/components/schemas/Timestamp'
        date_end_fact:
          $ref: '#/components/schemas/Timestamp'
        status_id:
          type: integer
...
    Timestamp:
      type: object
      required:
        - date
        - day
        - hours
        - minutes
        - month
        - seconds
        - time
        - timezoneOffset
        - year
      properties:
        date:
          type: integer
          format: int32
        day:
          type: integer
          format: int32
        hours:
          type: integer
          format: int32
        minutes:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
        seconds:
          type: integer
          format: int32
        time:
          type: integer
          format: int64
        timezoneOffset:
          type: integer
          format: int32
        year:
          type: integer
          format: int32
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string